Split Config into utils
