+++ src/utils/config/get-default-config.js
@@ -1,0 +1,3 @@
+import { defaultConfig } from '@utils/config';
+
+export const getDefaultConfig = () => defaultConfig;
